refactor: isolate app control and OpenAI realtime driver
Reorganize the realtime voice-control PR so the generic OpenWork control surface lives independently from the OpenAI Realtime driver. Move session-owned control actions into the session domain, move the OpenAI browser driver and activity/status UI into a driver folder, and move backend Realtime session/tool setup out of server.ts.
